in tests, pause `confirm` plugin instead of using `delay(5)`
When the `SELECT` key is sent to the `confirm` prompt, the intercepted key isn't passed on to the underlying screen until the next rendered frame. Since we escape out of that screen before the next rendered frame, it is never sent -- until the screen is reopened. If you run the test, and it fails, and then you manually open the manager orders screen, you can see the test order there before it quickly disappears. If you pause the confirm dialog and resend `MANAGER_REMOVE`, it works.

Co-authored-by: Myk <myk002@yahoo.com>